Completed upgrading to typescript
