+++ package.json
@@ -43,9 +43,13 @@
     ]
   },
   "devDependencies": {
+    "@types/node": "^20.4.0",
+    "@types/react": "^18.2.14",
+    "@types/react-dom": "^18.2.6",
     "autoprefixer": "^10.4.14",
     "postcss": "^8.4.24",
     "tailwind-scrollbar": "^2.1.0-preview.0",
-    "tailwindcss": "^3.3.2"
+    "tailwindcss": "^3.3.2",
+    "typescript": "^5.1.6"
   }
 }
