+++ src/components/Chart.tsx
@@ -1,70 +1,16 @@
 import React, { useLayoutEffect, useState } from "react";
-
-import {
-  LineChart,
-  Line,
-  XAxis,
-  YAxis,
-  Tooltip,
-  CartesianGrid,
-  Legend,
-  ResponsiveContainer,
-} from "recharts";
 import { useContext } from "react";
-import { CryptoContext } from "./../context/CryptoContext";
+import { CryptoContext } from "../context/CryptoContext";
+import { ChartComponent } from "./ChartComponent";
 
-function CustomTooltip({ payload, label, active, currency = "usd" }) {
-  if (active && payload && payload.length > 0) {
-    return (
-      <div className="custom-tooltip">
-        <p className="label text-sm text-cyan">{`${label} : ${new Intl.NumberFormat(
-          "en-IN",
-          {
-            style: "currency",
-            currency: currency,
-            minimumFractionDigits: 5,
-          }
-        ).format(payload[0].value)}`}</p>
-      </div>
-    );
-  }
-
-  return null;
-}
-
-const ChartComponent = ({ data, currency, type }) => {
-  return (
-    <ResponsiveContainer height={"90%"}>
-      <LineChart width={400} height={400} data={data}>
-        <Line
-          type="monotone"
-          dataKey={type}
-          stroke="#14ffec"
-          strokeWidth={"1px"}
-        />
-        <CartesianGrid stroke="#323232" />
-        <XAxis dataKey="date" hide />
-        <YAxis dataKey={type} hide domain={["auto", "auto"]} />
-        <Tooltip
-          content={<CustomTooltip />}
-          currency={currency}
-          cursor={false}
-          wrapperStyle={{ outline: "none" }}
-        />
-        <Legend />
-      </LineChart>
-    </ResponsiveContainer>
-  );
-};
-
-const Chart = ({ id }) => {
+const Chart = ({ id }: { id: any }) => {
   const [chartData, setChartData] = useState();
   let { currency } = useContext(CryptoContext);
   const [type, setType] = useState("prices");
   const [days, setDays] = useState(7);
 
   useLayoutEffect(() => {
-    const getChartData = async (id) => {
+    const getChartData = async (id: string) => {
       try {
         const data = await fetch(
           `https://api.coingecko.com/api/v3/coins/${id}/market_chart?vs_currency=usd&days=${days}&interval=daily`
@@ -74,7 +20,7 @@ const Chart = ({ id }) => {
 
         console.log("chart-data", data);
 
-        let convertedData = data[type].map((item) => {
+        let convertedData = data[type].map((item: any) => {
           return {
             date: new Date(item[0]).toLocaleDateString(),
             [type]: item[1],
