+++ src/components/ChartComponent.tsx
@@ -1,0 +1,36 @@
+import {
+  ResponsiveContainer,
+  LineChart,
+  Line,
+  CartesianGrid,
+  XAxis,
+  YAxis,
+  Tooltip,
+  Legend,
+} from "recharts";
+import { CustomTooltip } from "./CustomTooltip";
+
+export const ChartComponent = ({ data, currency, type }: any) => {
+  return (
+    <ResponsiveContainer height={"90%"}>
+      <LineChart width={400} height={400} data={data}>
+        <Line
+          type="monotone"
+          dataKey={type}
+          stroke="#14ffec"
+          strokeWidth={"1px"}
+        />
+        <CartesianGrid stroke="#323232" />
+        <XAxis dataKey="date" hide />
+        <YAxis dataKey={type} hide domain={["auto", "auto"]} />
+        <Tooltip
+          content={<CustomTooltip />}
+          //   currency={currency}
+          cursor={false}
+          wrapperStyle={{ outline: "none" }}
+        />
+        <Legend />
+      </LineChart>
+    </ResponsiveContainer>
+  );
+};
