+++ src/components/CustomTooltip.tsx
@@ -1,0 +1,23 @@
+export function CustomTooltip({
+  payload,
+  label,
+  active,
+  currency = "usd",
+}: any) {
+  if (active && payload && payload.length > 0) {
+    return (
+      <div className="custom-tooltip">
+        <p className="label text-sm text-cyan">{`${label} : ${new Intl.NumberFormat(
+          "en-IN",
+          {
+            style: "currency",
+            currency: currency,
+            minimumFractionDigits: 5,
+          }
+        ).format(payload[0].value)}`}</p>
+      </div>
+    );
+  }
+
+  return null;
+}
