+++ src/components/HighLowIndicator.tsx
@@ -1,0 +1,28 @@
+import { useState, useEffect } from "react";
+
+export const HighLowIndicator = ({ currentPrice, high, low }: any) => {
+  const [green, setGreen] = useState<number>(0);
+
+  useEffect(() => {
+    let total = high - low;
+    let greenZone = ((high - currentPrice) * 100) / total;
+    setGreen(Math.ceil(greenZone));
+  }, [currentPrice, high, low]);
+
+  return (
+    <>
+      <span
+        className="bg-red h-1.5 rounded-l-lg w-[50%]"
+        style={{ width: `${100 - green}%` }}
+      >
+        &nbsp;
+      </span>
+      <span
+        className="bg-green h-1.5 rounded-r-lg w-[50%]"
+        style={{ width: `${green}%` }}
+      >
+        &nbsp;
+      </span>
+    </>
+  );
+};
