+++ src/components/Pagination.tsx
@@ -1,63 +1,19 @@
-import React, { useContext, useRef } from "react";
+import React, { useContext } from "react";
 import paginationArrow from "../assets/pagination-arrow.svg";
-import { CryptoContext } from "./../context/CryptoContext";
-import submitIcon from "../assets/submit-icon.svg";
-
-const PerPage = () => {
-  const { setPerPage } = useContext(CryptoContext);
-  const inputRef = useRef(null);
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    let val = inputRef.current.value;
-    if (val !== 0) {
-      setPerPage(val);
-      inputRef.current.value = val;
-    }
-  };
-
-  return (
-    <form
-      className="relative flex items-center font-nunito
-        mr-12
-        "
-      onSubmit={handleSubmit}
-    >
-      <label
-        htmlFor="perpage"
-        className="relative flex justify-center items-center
-        mr-2 font-bold
-        "
-      >
-        per page:{" "}
-      </label>
-      <input
-        type="number"
-        name="perpage"
-        min={1}
-        max={250}
-        ref={inputRef}
-        placeholder="10"
-        className="w-16 rounded bg-gray-200 placeholder:text-gray-100
-    pl-2 required outline-0 border border-transparent 
-    focus:border-cyan leading-4
-    "
-      />
-      <button type="submit" className="ml-1 cursor-pointer">
-        <img src={submitIcon} alt="submit" className="w-full h-auto" />
-      </button>
-    </form>
-  );
-};
+import { CryptoContext } from "../context/CryptoContext";
+import useScreenSize from "../hooks/useScreenSize";
+import { PerPage } from "./PerPage";
 
 const Pagination = () => {
   let { page, setPage, totalPages, perPage, cryptoData } =
     useContext(CryptoContext);
 
-  const TotalNumber = Math.ceil(totalPages / perPage);
+  const screenSize = useScreenSize();
+
+  const totalNumber = Math.ceil(totalPages / perPage);
 
   const next = () => {
-    if (page === TotalNumber) {
+    if (page === totalNumber) {
       return null;
     } else {
       setPage(page + 1);
@@ -73,8 +29,8 @@ const Pagination = () => {
   };
 
   const multiStepNext = () => {
-    if (page + 3 >= TotalNumber) {
-      setPage(TotalNumber - 1);
+    if (page + 3 >= totalNumber) {
+      setPage(totalNumber - 1);
     } else {
       setPage(page + 3);
     }
@@ -82,7 +38,7 @@ const Pagination = () => {
 
   const multiStepPrev = () => {
     if (page - 3 <= 1) {
-      setPage(TotalNumber + 1);
+      setPage(totalNumber + 1);
     } else {
       setPage(page - 2);
     }
@@ -90,7 +46,11 @@ const Pagination = () => {
 
   if (cryptoData && cryptoData.length >= perPage) {
     return (
-      <div className="flex items-center flex-1 w-full md:w-auto justify-between my-3 md:my-0">
+      <div
+        className={`flex items-center flex-1 w-full md:w-auto justify-between my-3 md:my-0 ${
+          screenSize === "xs" ? "flex-col" : "flex-row"
+        }`}
+      >
         <PerPage />
         <ul className="flex items-center justify-end text-sm">
           <li className="flex items-center">
@@ -103,7 +63,7 @@ const Pagination = () => {
             </button>
           </li>
 
-          {page + 1 === TotalNumber || page === TotalNumber ? (
+          {page + 1 === totalNumber || page === totalNumber ? (
             <li>
               {" "}
               <button
@@ -135,7 +95,7 @@ const Pagination = () => {
             </button>
           </li>
 
-          {page + 1 !== TotalNumber && page !== TotalNumber ? (
+          {page + 1 !== totalNumber && page !== totalNumber ? (
             <li>
               <button
                 onClick={next}
@@ -146,7 +106,7 @@ const Pagination = () => {
             </li>
           ) : null}
 
-          {page + 1 !== TotalNumber && page !== TotalNumber ? (
+          {page + 1 !== totalNumber && page !== totalNumber ? (
             <li>
               {" "}
               <button
@@ -158,13 +118,13 @@ const Pagination = () => {
             </li>
           ) : null}
 
-          {page !== TotalNumber ? (
+          {page !== totalNumber ? (
             <li>
               <button
-                onClick={() => setPage(TotalNumber)}
+                onClick={() => setPage(totalNumber)}
                 className="ouline-0 hover:text-cyan  rounded-full w-fit p-1 h-8 flex items-center justify-center bg-gray-200 mx-1.5"
               >
-                {TotalNumber}
+                {totalNumber}
               </button>
             </li>
           ) : null}
