+++ src/pages/Trending.tsx
@@ -11,7 +11,7 @@ const Trending = () => {
     <section className="w-11/12 sm:w-[80%] h-full flex flex-col mt-16 mb-24 relative">
       <div className="w-full min-h-[60vh] py-8 flex flex-wrap justify-evenly  border border-gray-100 rounded">
         {trendData &&
-          trendData.map((coin) => (
+          trendData.map((coin: any) => (
             <TrendingCoin key={data.coin_id} data={coin.item} />
           ))}
         <button
